% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jptransnetdata-package.R
\docType{package}
\name{jptransnetdata-package}
\alias{jptransnetdata}
\alias{jptransnetdata-package}
\title{jptransnetdata: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
